refactor : repository -> domain으로 분리 후 controller에서 비즈니스 로직 분리
