+++ src/main/java/domain/ResultCalculator.java
@@ -1,29 +1,26 @@
-package repository;
+package domain;
 
-import domain.Lotto;
-import domain.Rank;
-import domain.WinningNumber;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-public class LottoResultRepository {
+public class ResultCalculator {
 
     private final Map<Rank, Integer> calculateResult;
 
-    private LottoResultRepository() {
+    private ResultCalculator() {
         this.calculateResult = new LinkedHashMap<>();
         for (Rank value : Rank.values()) {
             calculateResult.put(value, 0);
         }
     }
 
-    public static LottoResultRepository create() {
-        return new LottoResultRepository();
+    public static ResultCalculator create() {
+        return new ResultCalculator();
     }
 
-    public void add(WinningNumber winningNumber, List<Lotto> lottos) {
+    public void calculate(WinningNumber winningNumber, List<Lotto> lottos) {
         for (Lotto lotto : lottos) {
             int matchCount = winningNumber.calculateMatchNumber(lotto);
             boolean hasBonusNumber = winningNumber.hasBonusNumber(lotto);
